requirements.txt
discord.py==2.4.0
PyYAML==6.0.3
